backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP and API clients
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Google Sheets integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# AI and ML
langchain==0.1.0
openai==1.3.7
anthropic==0.7.8

# Email
smtplib-ssl==1.0.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Background tasks
celery==5.3.4
redis==5.0.1

# File processing
PyPDF2==3.0.1
python-magic==0.4.27

# Date and time
python-dateutil==2.8.2

# Utilities
click==8.1.7
rich==13.7.0
